:sparkles: 인스턴스 authInstance, baseInstance로 로직 분리
